More scrolling/event pass through fixes. Move view valid check to renderer
